Pagamento2: cambia to il tipo della variabile "quant"

+++ Classes/Pagamento2.h
@@ -21,7 +21,7 @@
 	
     NSInteger identificativo;
 	double valore;
-	double quant;
+	int quant;
 	double totale;
 	
     IBOutlet UIButton *compra;
